pyproject.toml
[project]
name = "lorelm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    # web
    "fastapi>0.100",
    "python-multipart",
    "uvicorn",
    "uvloop; platform_system == 'Darwin' or platform_system == 'Linux'",
    "gunicorn>=21.2.0",
    "pydantic>2,<3",
    "pyjwt>=2",
    "cryptography>45",
    "pydantic[email]",
    "orjson>=3",
    # database
    "alembic",
    "sqlalchemy[asyncio]",
    # mysql
    # "aiomysql; platform_system == 'Windows'",
    # "asyncmy; platform_system == 'Darwin' or platform_system == 'Linux'",
    # postgresql
    "psycopg>3",
    "pgvector>0.4",
    # large language model
    "json-repair",
    "openai",
    "omni-llm",
    "jinja2",
    "tiktoken",
    # rag
    "numpy",
    "jieba>=0.42.1",
    "tokenizers",
    "modelscope",
    "hanziconv",
    "nltk",
    "datrie==0.8.3",
    "markdown>=3.8.2",
    "elasticsearch[async]==8.18.1",
    # system
    "coloredlogs",
    "python-dotenv",
    "miniopy-async",
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true


[tool.uv.sources]
omni-llm = { index = "omni-llm" }

[[tool.uv.index]]
name = "omni-llm"
url = "http://zkr-ora.gitlab.io:11080/api/v4/projects/5/packages/pypi/simple"
explicit = true
authenticate = "always"
